Update wbsrc.sh script

+++ wbsrc.sh
@@ -2,9 +2,12 @@
 # WB=filename.NEF - white balance reference file
 # NOLG=1 - do not use lukaszgryglicki's "jpeg/jpegbw" tools - use ImageMagick's "convert" instead
 # MULT="R G B G" - manually specify white balance, example: 0.37033 1.0 0.525732 0.996594, overrides WB
-# DCARGS="-v -H 0" - manually specify dcraw args for conversion, default is "-H 1 -W -q 3 -6", other can be "-v -H 0 -q 0 -b 1.2 -g 2.1 4.2" for example
+# DCARGS="-v -H 1" - manually specify dcraw args for conversion, default is "-v -H 0 -W -q 3 -6", other can be "-H 1 -q 0 -b 1.2 -g 2.1 4.2" for example
 # The most imporant dcraw param is: -H [0-9] Highlight mode (0=clip, 1=unclip, 2=blend, 3+=rebuild)
 # -W Don't automatically brighten the image
+# LO=0-100 - set LO enhance cutoff - default 2 (percent)
+# HI=0-100 - set HI enhance cutoff - default 0.5 (percent)
+# NODEL=1 - do not remove intermediate files
 if [ -z "${MULT}" ]
 then
   if [ -z "${WB}" ]
@@ -18,7 +21,10 @@ then
     ppm="${WB/.NEF/.ppm}"
     echo "generating white balance info: ${inf}"
     dcraw -v -H 1 -W -a "${WB}" 2> "${inf}" || exit 2
-    rm -rf "${ppm}"
+    if [ -z "$NODEL" ]
+    then
+      rm -rf "${ppm}"
+    fi
   fi
   mult=`cat "${inf}" | grep multipliers`
   mult="${mult/multipliers /}"
@@ -26,9 +32,17 @@ then
 fi
 if [ -z "${DCARGS}" ]
 then
-  export DCARGS="-H 1 -W -q 3 -6"
+  export DCARGS="-v -H 0 -W -q 3 -6"
 fi
-echo "using RGRB multipliers: ${MULT}, dcraw args: ${DCARGS}"
+if [ -z "$LO" ]
+then
+  export LO="2"
+fi
+if [ -z "$HI" ]
+then
+  export HI="0.2"
+fi
+echo "using RGRB multipliers: ${MULT}, dcraw args: ${DCARGS}, lo/hi=$LO/$HI"
 for f in "$@"
 do
   if [[ $f != *.NEF ]]
@@ -48,13 +62,22 @@ do
   if [ -z "${NOLG}" ]
   then
     convert "${tiff}" -enhance -quality 99% "${jpeg}" || exit 4
-    rm -rf "${tiff}"
-    RR=1 RG=0 RB=0 GR=0 GG=1 GB=0 BR=0 BG=0 BB=1 RLO=3 RHI=.05 GLO=3 GHI=.05 BLO=3 BHI=.05 NA=1 ACM=1 Q=95 jpeg "${jpeg}" || exit 5
-    rm -rf "${jpeg}"
+    if [ -z "$NODEL" ]
+    then
+      rm -rf "${tiff}"
+    fi
+    RR=1 RG=0 RB=0 GR=0 GG=1 GB=0 BR=0 BG=0 BB=1 RLO=$LO RHI=$HI GLO=$LO GHI=$HI BLO=$LO BHI=$HI NA=1 ACM=1 Q=96 jpeg "${jpeg}" || exit 5
+    if [ -z "$NODEL" ]
+    then
+      rm -rf "${jpeg}"
+    fi
     echo "saved co_${jpeg}"
   else
     convert "${tiff}" -auto-level -enhance -quality 95% "${jpeg}" || exit 6
-    rm -rf "${tiff}"
+    if [ -z "$NODEL" ]
+    then
+      rm -rf "${tiff}"
+    fi
     echo "saved ${jpeg}"
   fi
 done
